refactor(loyalty): isolate the transfer path instead of splitting the ledger
Reworks the previous attempt. The cohesion finding was a delta, not a standing
property of queries.go: that file already carried the ledger and the named
counters, and folding a third multi-step operation into it is what pushed it
under the bar. Splitting a pre-existing half back out only moved the problem,
since a newly created file is held to the health bar for new code and those
long-standing signatures were never written to it.

So queries.go returns to its baseline byte for byte, and only the new work
lands in transfer.go. Its arguments become a Transfer struct: channel, sender,
recipient and amount were otherwise four scalars in a row, two of them bare
uint64s that swap silently.

boundedAmount splits its parse failure from its range check, so neither branch
is a three-operand conditional.
